requirements.txt
flask>=3.0.0
dulwich>=0.22.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
